Several fixes for CSS support.
